% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{count_go_terms}
\alias{count_go_terms}
\title{Quantify GO terms based on data frame containing ensembl ids.}
\usage{
count_go_terms(go_frame, go_description)
}
\arguments{
\item{go_frame}{Frame with layout `colnames(c(ensembl_id, go_id))`}

\item{go_description}{Go description frame containing at least `colnames(go_id)`}
}
\value{
go_description frame with new `count` column containing the quantity of each unique `ensembl_id` in `go_id`
}
\description{
Requires a data frame in the following format
  `go_frame`: colnames(c(ensembl_id, go_id))
  `go_description`: colnames(go_id)
The function iterates over all elements in `go_frame` and counts how many elements are present for each `go_id`
}
