% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{get_current_biomart_version}
\alias{get_current_biomart_version}
\title{Get the current biomaRt version}
\usage{
get_current_biomart_version()
}
\value{
Biomart version
}
\description{
Get the current biomaRt version
}
