% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{get_current_ensembl_version}
\alias{get_current_ensembl_version}
\title{Get the current Ensembl version}
\usage{
get_current_ensembl_version()
}
\value{
Ensembl version
}
\description{
Get the current Ensembl version
}
