% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{get_ensembl_data}
\alias{get_ensembl_data}
\title{Get the ensembl data frame requested by 'type'.}
\usage{
get_ensembl_data(type = "gene_and_go",
  ensembl_dataset = "mmusculus_gene_ensembl", version = "current")
}
\arguments{
\item{type}{Ensembl data to return. Has to be `gene_and_go`, `transcripts_and_go` or `go_description`}

\item{ensembl_dataset}{Ensembl dataset name specifying the species}

\item{ensembl_version}{Ensembl version. Defaults to 'current'}
}
\value{
Dataframe reqested by type
}
\description{
Since the download requires some minutes, files are cached in the installation folder of the package.
Therefore, loading a gene dataset can take some time when done for the first time.
The data frames are stored in individual folders to make loading more efficient
}
