% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{get_go_per_gene}
\alias{get_go_per_gene}
\title{Export the gene_and_go table where each gene is mapped to a corresponding GO term}
\usage{
get_go_per_gene(ensembl_dataset = "mmusculus_gene_ensembl",
  ensembl_version = "current")
}
\arguments{
\item{ensembl_dataset}{Ensembl dataset name specifying the species}

\item{ensembl_version}{Ensembl version. Defaults to 'current'}
}
\description{
Export the gene_and_go table where each gene is mapped to a corresponding GO term
}
