% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{get_go_summary}
\alias{get_go_summary}
\title{Get summary table containing go terms accession id and description}
\usage{
get_go_summary(ensembl_dataset = "mmusculus_gene_ensembl",
  ensembl_version = "current")
}
\arguments{
\item{ensembl_dataset}{Ensembl dataset name specifying the species}

\item{ensembl_version}{Ensembl version. Defaults to 'current'}
}
\value{
Summary table
}
\description{
Get summary table containing go terms accession id and description
}
