% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{to_go}
\alias{to_go}
\title{Get the GO term association per gene or transcript as list}
\usage{
to_go(input)
}
\arguments{
\item{input}{Vector of Ensembl gene or transcript ids}
}
\value{
List named by gene/transcript id containing a vector of associated GO terms by accession id
}
